refactor: :art: Organize code in packages
