+++ car/interfaces.go
@@ -1,0 +1,10 @@
+package car
+
+import "CarSim/parts"
+
+type CarManipulator interface {
+	WheelManiputlator() parts.WheelsManipulator
+	DoorManipulator() parts.DoorManipulator
+	MotorManipulator() parts.MotorManipulator
+	Print()
+}
