+++ constant/constant.go
@@ -1,0 +1,6 @@
+package constant
+
+const (
+	WheelCount       = 4
+	DefaultWheelSize = 20
+)
